requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncclick==8.3.0.3
asyncio==4.0.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
dataclass-csv==1.4.1
exceptiongroup==1.3.1
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-kasa==0.7.7
requests==2.33.1
six==1.17.0
slack==0.0.2
suntime==1.3.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
